Update to be independent of location

+++ pyxb/bundles/opengis/examples/test.sh
@@ -1,10 +1,4 @@
-PYXB_ROOT=${PYXB_ROOT:-/home/pab/pyxb/dev}
-BUNDLE_ROOT=${PYXB_ROOT}/pyxb/bundles/opengis
-SCHEMA_DIR=${BUNDLE_ROOT}/schemas
-
-PYTHONPATH=${PYXB_ROOT}:.
-PATH=${PYXB_ROOT}/scripts:/usr/bin:/bin
-export PATH PYTHONPATH
+#!/bin/sh
 
 failure () {
   echo "Failed: ${@}"
@@ -15,14 +9,16 @@ python demo.py || exit 1
 
 # sosRegisterSensor.xml uses tml:tcfTrigger, but the element is really named tml:cfTrigger
 # Skip testing it since it will fail to validate thereby confusing the viewer.
-ls ${SCHEMA_DIR}/sos/1.0.0/examples/*.xml \
- | grep -v sosRegisterSensor \
- | xargs python check_sos.py \
-|| exit 1
+if [ -d ${SCHEMAS_OPENGIS_NET:-/dev/null} ] ; then
+  ls ${SCHEMAS_OPENGIS_NET}/sos/1.0.0/examples/*.xml \
+    | grep -v sosRegisterSensor \
+    | xargs python check_sos.py \
+  || exit 1
+fi
 
 rm -f gmlapp.py raw/gmlapp.py
 pyxbgen \
-  --archive-path=${BUNDLE_ROOT}/raw \
   --schema-location=gmlapp.xsd --module=gmlapp \
   --write-for-customization
-python testgml.py
+python testgml.py || exit 1
+
